refactor(dashboard): update dashboard UI components to import types from src/types/dashboard
